+++ src/full_text/FullTextResultSet.cpp
@@ -26,7 +26,7 @@ void FullTextResultSet::allocate(size_t len) {
 
 	if (len > 0) {
 		m_value_pointer = new uint64_t[len];
-		m_score_pointer = new uint32_t[len];
+		m_score_pointer = new float[len];
 		m_has_allocated = true;
 		m_len = len;
 	}
@@ -37,7 +37,7 @@ uint64_t *FullTextResultSet::value_pointer() {
 	return m_value_pointer;
 }
 
-uint32_t *FullTextResultSet::score_pointer() {
+float *FullTextResultSet::score_pointer() {
 	return m_score_pointer;
 }
