+++ src/link_index/LinkResultSet.h
@@ -21,7 +21,7 @@ public:
 	uint64_t *target_pointer();
 	uint64_t *source_domain_pointer();
 	uint64_t *target_domain_pointer();
-	uint32_t *score_pointer();
+	float *score_pointer();
 	void set_total_num_results(size_t total_num_results);
 
 private:
@@ -31,7 +31,7 @@ private:
 	uint64_t *m_target_pointer;
 	uint64_t *m_source_domain_pointer;
 	uint64_t *m_target_domain_pointer;
-	uint32_t *m_score_pointer;
+	float *m_score_pointer;
 	bool m_has_allocated;
 	size_t m_len;
 	size_t m_total_num_results;
